+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/MyCookieManager.java
@@ -171,13 +171,16 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
       });
       cookieManager.flush();
     }
-    else {
+    else if (plugin != null) {
       CookieSyncManager cookieSyncMngr = CookieSyncManager.createInstance(plugin.applicationContext);
       cookieSyncMngr.startSync();
       cookieManager.setCookie(url, cookieValue);
       result.success(true);
       cookieSyncMngr.stopSync();
       cookieSyncMngr.sync();
+    } else {
+      cookieManager.setCookie(url, cookieValue);
+      result.success(true);
     }
   }
 
@@ -218,7 +221,12 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
     cookieManager = getCookieManager();
     if (cookieManager == null) return;
 
-    String cookieValue = name + "=; Path=" + path + "; Domain=" + domain + "; Max-Age=-1;";
+    String cookieValue = name + "=; Path=" + path + "; Max-Age=-1";
+
+    if (domain != null)
+      cookieValue += "; Domain=" + domain;
+
+    cookieValue += ";";
 
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
       cookieManager.setCookie(url, cookieValue, new ValueCallback<Boolean>() {
@@ -229,13 +237,16 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
       });
       cookieManager.flush();
     }
-    else {
+    else if (plugin != null) {
       CookieSyncManager cookieSyncMngr = CookieSyncManager.createInstance(plugin.applicationContext);
       cookieSyncMngr.startSync();
       cookieManager.setCookie(url, cookieValue);
       result.success(true);
       cookieSyncMngr.stopSync();
       cookieSyncMngr.sync();
+    } else {
+      cookieManager.setCookie(url, cookieValue);
+      result.success(true);
     }
   }
 
@@ -248,7 +259,7 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
     String cookiesString = cookieManager.getCookie(url);
     if (cookiesString != null) {
 
-      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
+      if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP && plugin != null) {
         cookieSyncMngr = CookieSyncManager.createInstance(plugin.applicationContext);
         cookieSyncMngr.startSync();
       }
@@ -257,7 +268,14 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
       for (String cookie : cookies) {
         String[] nameValue = cookie.split("=", 2);
         String name = nameValue[0].trim();
-        String cookieValue = name + "=; Path=" + path + "; Domain=" + domain + "; Max-Age=-1;";
+
+        String cookieValue = name + "=; Path=" + path + "; Max-Age=-1";
+
+        if (domain != null)
+          cookieValue += "; Domain=" + domain;
+
+        cookieValue += ";";
+
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
           cookieManager.setCookie(url, cookieValue, null);
         else
@@ -286,13 +304,16 @@ public class MyCookieManager implements MethodChannel.MethodCallHandler {
       });
       cookieManager.flush();
     }
-    else {
+    else if (plugin != null) {
       CookieSyncManager cookieSyncMngr = CookieSyncManager.createInstance(plugin.applicationContext);
       cookieSyncMngr.startSync();
       cookieManager.removeAllCookie();
       result.success(true);
       cookieSyncMngr.stopSync();
       cookieSyncMngr.sync();
+    } else {
+      cookieManager.removeAllCookie();
+      result.success(true);
     }
   }